requirements.txt
google-cloud-aiplatform>=1.62.0,<2.0.0
google-cloud-secret-manager>=2.20.0,<3.0.0
fastapi>=0.113.0,<1.0.0
uvicorn[standard]>=0.30.1,<1.0.0
pydantic>=2.8.2,<3.0.0
Pillow>=10.4.0,<11.0.0
PyMuPDF>=1.24.8,<1.25.0
google-genai>=1.16.1 